requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pytest
pytest-asyncio
httpx
jinja2
pydantic[email]
passlib==1.7.4
python-multipart
openai
transformers>=4.40
torch>=2.0
pytest-asyncio
